refactor: move state module into state/ package
Phase 5 of #625 - reorganize project structure:
- state_machine.py → state/machine.py
- state_reader.py → state/reader.py
- transition_validator.py → state/validator.py
- Tests moved to tests/state/
- Updated all imports and mock patches
