+++ custom_components/localshift/state/reader.py
@@ -10,7 +10,7 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.util import dt as dt_util
 
-from .const import (
+from ..const import (
     CONF_PRICING_FEED_IN_FORECAST,
     CONF_PRICING_FEED_IN_PRICE,
     CONF_PRICING_GENERAL_FORECAST,
@@ -29,8 +29,8 @@ from .const import (
     CONF_WEATHER_ENTITY,
     DEFAULT_ENTITY_IDS,
 )
-from .coordinator_data import CoordinatorData
-from .utils.validation import EntityValidator
+from ..coordinator_data import CoordinatorData
+from ..utils.validation import EntityValidator
 
 _LOGGER = logging.getLogger(__name__)
